+++ DMGwGB/DMGwGB/CRule.h
@@ -10,6 +10,6 @@ public:
 
 	int grain_count = 0;
 
-	virtual void step(unsigned short * cell, class std::vector<unsigned short> neighborhood);
+	virtual void step(unsigned int * cell, class std::vector<unsigned int> neighborhood);
 };
