+++ DMGwGB/DMGwGB/GrainGrowthRule.h
@@ -7,6 +7,6 @@ public:
 	GrainGrowthRule();
 	~GrainGrowthRule();
 	
-	virtual void step(unsigned short * cell, class std::vector<unsigned short > neighborhood);
+	virtual void step(unsigned int * cell, class std::vector<unsigned int > neighborhood);
 };
