+++ DMGwGB/DMGwGB/VonNeummanNeighborhood.h
@@ -7,6 +7,6 @@ public:
 	VonNeummanNeighborhood();
 	~VonNeummanNeighborhood();
 
-	virtual std::vector<unsigned short> get(class CellularAutomata * cellularautomata, unsigned int x, unsigned int y, unsigned int z);
+	virtual std::vector<unsigned int> get(class CellularAutomata * cellularautomata, unsigned int x, unsigned int y, unsigned int z);
 };
